Setters for NSUserNotification
While writing the setters I started to think about how/what will be moved into a separate file. I realized that once most of the Obj-C stuff is moved that it would make more sense to keep it all in one place and have the file act as an abstraction layer.  This setup would allow the NS/CF related functionality to sit in a library that can be easily called from C.